+++ NEWS.md
@@ -41,7 +41,7 @@
   * `vec_group_rle()`
   * `vec_locate_matches()`
 
-* Hashing is now supported for lists containing complex vectors, enabling functions like `vec_unique_loc()` to work on these objects (#1992).
+* Hashing is now supported for lists containing complex or raw vectors, enabling functions like `vec_unique_loc()` to work on these objects (#1992, #2046).
 
 * `vec_detect_complete(NULL)` now returns `logical()`, consistent with
   `vec_detect_missing(NULL)` (#1916).